+++ openlane/scripts/openroad/gpl.tcl
@@ -73,7 +73,7 @@ set cell_pad_side [expr $::env(GPL_CELL_PADDING) / 2]
 lappend arg_list -pad_right $cell_pad_side
 lappend arg_list -pad_left $cell_pad_side
 lappend arg_list -init_wirelength_coef $::env(PL_WIRE_LENGTH_COEF)
-lappend arg_list -keep_resize_below_overflow $::env(PL_KEEP_RESIZE_BELOW_OVERFLOW)
+append_if_exists_argument arg_list PL_KEEP_RESIZE_BELOW_OVERFLOW -keep_resize_below_overflow
 
 log_cmd global_placement {*}$arg_list