+++ openlane/steps/openroad.py
@@ -1290,9 +1290,8 @@ class _GlobalPlacement(OpenROADStep):
             ),
             Variable(
                 "PL_KEEP_RESIZE_BELOW_OVERFLOW",
-                Decimal,
-                "Only applicable when PL_TIME_DRIVEN is enabled. When the overflow is below the set value, timing-driven iterations will retain the resizer changes instead of reverting them.",
-                default=0.5,
+                Optional[Decimal],
+                "Only applicable when PL_TIME_DRIVEN is enabled. When the overflow is below the set value, timing-driven iterations will retain the resizer changes instead of reverting them. Allowed values are 0 to 1.",
             ),
         ]
     )